aijournal-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib
bcrypt==4.0.1
requests
python-dotenv
python-multipart